telegram_bot/requirements.txt
telepot
requests
matplotlib
pytz
python-dotenv